+++ tests/serde-issues.rs
@@ -709,3 +709,50 @@ fn issue888() {
         }
     );
 }
+
+/// Regression test for https://github.com/tafia/quick-xml/issues/906.
+#[test]
+fn issue906() {
+    #[derive(Debug, PartialEq, Deserialize, Serialize)]
+    struct AsElement {
+        #[serde(rename = "a-list")]
+        a_list: TextContent,
+    }
+
+    #[derive(Debug, PartialEq, Deserialize, Serialize)]
+    struct AsAttribute {
+        #[serde(rename = "@a-list")]
+        a_list: TextContent,
+    }
+
+    #[derive(Debug, PartialEq, Deserialize, Serialize)]
+    struct TextContent {
+        #[serde(rename = "$text")]
+        content: Vec<String>,
+        #[serde(skip)]
+        ignored: (),
+    }
+
+    let foo = AsElement {
+        a_list: TextContent {
+            content: vec!["A".to_string(), "B".to_string()],
+            ignored: (),
+        },
+    };
+    let bar = AsAttribute {
+        a_list: TextContent {
+            content: vec!["A".to_string(), "B".to_string()],
+            ignored: (),
+        },
+    };
+
+    let buffer = to_string_with_root("test", &foo).unwrap();
+    assert_eq!(buffer, "<test><a-list>A B</a-list></test>");
+    let foo2: AsElement = from_str(&buffer).unwrap();
+    assert_eq!(foo2, foo);
+
+    let buffer = to_string_with_root("test", &bar).unwrap();
+    assert_eq!(buffer, "<test a-list=\"A B\"/>");
+    let bar2: AsAttribute = from_str(&buffer).unwrap();
+    assert_eq!(bar2, bar);
+}